+++ src/ffi3/object.rs
@@ -564,9 +564,9 @@ mod typeobject {
     pub unsafe fn PyHeapType_GET_MEMBERS(
         etype: *mut PyHeapTypeObject,
     ) -> *mut ffi3::structmember::PyMemberDef {
-        (etype as *mut c_char).offset(
-            (*ffi3::object::Py_TYPE(etype as *mut ffi3::object::PyObject)).tp_basicsize as isize,
-        ) as *mut ffi3::structmember::PyMemberDef
+        let py_type = ffi3::object::Py_TYPE(etype as *mut ffi3::object::PyObject);
+        let ptr = (etype as *mut c_char).offset((*py_type).tp_basicsize);
+        ptr as *mut ffi3::structmember::PyMemberDef
     }
 }